VER-80: a functioning version of all pages as per figma.
